requirements.txt
aniso8601==7.0.0
asgiref==3.2.10
astroid==2.4.2
autopep8==1.5.3
Django==3.0.8
django-currentuser==0.5.0
django-filter==2.3.0
django-graphql-jwt==0.3.1
django-rest-auth==0.9.5
django-rest-framework==0.1.0
django-webserver==1.2.0
djangorestframework==3.11.0
graphene==2.1.8
graphene-django==2.11.1
graphene-django-extras==0.4.9
graphene-django-optimizer==0.6.2
graphql-core==2.3.2
graphql-relay==2.0.1
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
promise==2.3
pycodestyle==2.6.0
PyJWT==1.7.1
pylint==2.5.3
pylint-django==2.0.15
pylint-plugin-utils==0.6
python-dateutil==2.8.1
pytz==2020.1
Rx==1.6.1
singledispatch==3.4.0.3
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
Unidecode==1.1.1
wrapt==1.12.1
